pyproject.toml
[project]
name = "docopt-typed"
version = "0.1.0"
description = "A lightweight helper around docopt-ng adding automatic type conversion"
readme = "README.md"
authors = [
    { name = "John Scillieri" }
]
requires-python = ">=3.13"
dependencies = [
    "docopt-ng>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.12.10",
]

[tool.ruff]
line-length = 120
target-version = "py313"
select = [
    "ALL", # include all the rules, including new ones
]
ignore = [
    "D400",
    "D415",
    "T201",
]
